Refactoring Endpoints
