+++ .gitignore
@@ -1,3 +1,5 @@
 /node_modules
 /public/bower_components
 /.idea
+Endpoints_Old.txt
+Endpoints_New.txt
