+++ public/services/jugend.js
@@ -1,37 +1,38 @@
 angular
     .module('spi.jugend', ['spi.auth']).factory('jugend', ['$http', 'auth', function ($http) {
 
+    var ENDPOINT_URL = '/jugenden';
     var jugend = {
         jugenden: []
     };
 
     jugend.getAll = function () {
-        return $http.get('/jugenden').success(function (data) {
+        return $http.get(ENDPOINT_URL).success(function (data) {
             angular.copy(data, jugend.jugenden);
             return data;
         });
     };
 
+    jugend.get = function (id) {
+        return $http.get(ENDPOINT_URL + '?id=' + id).then(function (res) {
+            return _.head(res.data);
+        });
+    };
+
     jugend.create = function (newjugend) {
-        return $http.post('/jugenden', newjugend).success(function (data) {
+        return $http.post(ENDPOINT_URL, newjugend).success(function (data) {
             return data;
         });
     };
 
-    jugend.get = function (id) {
-        return $http.get('/jugenden/' + id).then(function (res) {
-            return res.data;
-        });
-    };
-
     jugend.delete = function (id) {
-        return $http.delete('/jugenden/' + id).then(function (res) {
+        return $http.delete(ENDPOINT_URL + '?id=' + id).then(function (res) {
             return res;
         })
     };
 
     jugend.update = function (jugendId, jugend) {
-        return $http.put('/jugenden/' + jugendId, jugend).then(function (res) {
+        return $http.put(ENDPOINT_URL + '?id=' + jugendId, jugend).then(function (res) {
             return res.data;
         })
     };
@@ -43,7 +44,13 @@ angular
     };
 
     jugend.getTore = function (id) {
-        return $http.get('/jugenden/' + id + '/tore').then(function (res) {
+        return $http.get(ENDPOINT_URL + '/tore?id=' + id).then(function (res) {
+            return res;
+        });
+    };
+
+    jugend.getGesamtTore = function () {
+        return $http.get(ENDPOINT_URL + '/tore').then(function (res) {
             return res;
         });
     };
